requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
av
backoff
bcrypt
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
coloredlogs
contourpy
ctranslate2
cycler
dataclasses-json
decorator
Deprecated
distro
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
durationpy
faiss-cpu
faster-whisper
filelock
flatbuffers
fonttools
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
gunicorn
h11
hf-xet
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
jax
jaxlib
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mediapipe
ml-dtypes
mmh3
moviepy
mpmath
multidict
mypy-extensions
networkx
nltk
numpy
oauthlib
onnxruntime
opencv-contrib-python
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opt-einsum
orjson
ormsgpack
overrides
packaging
pandas
pillow
posthog
proglog
propcache
protobuf
pyasn1
pyasn1-modules
pybase64
pycparser
pydantic
pydantic-settings
Pygments
PyJWT
pyparsing
PyPika
pyproject-hooks
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools
shellingham
six
sounddevice
SQLAlchemy
sqlparse
sympy
tenacity
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
typer
typing-extensions
typing-inspect
urllib3
uuid-utils
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
xgboost
xxhash
yarl
zipp
zstandard
